+++ src/os2datascanner/projects/report/organizations/migrations/0028_account_handled_matches.py
@@ -6,7 +6,7 @@ from django.db import migrations, models
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('organizations', '0025_support_button_settings'),
+        ('organizations', '0027_organization_msgraph_write_permissions'),
     ]
 
     operations = [